requirements.txt
fastapi==0.115.13
uvicorn==0.34.3
sentence-transformers==4.1.0
faiss-cpu==1.11.0
transformers==4.52.4
torch==2.7.1
requests==2.32.4
pydantic==2.11.7
huggingface-hub==0.33.0

